Focus-Buddy/requirements-api.txt
fastapi==0.115.12
uvicorn==0.34.2
requests
python-dotenv
